Sales Module

+++ src/Shared/Shared.Application/Dtos/SharedCartItemDto.cs
@@ -5,6 +5,6 @@ public class SharedCartItemDto
     public Guid ProductId { get; set; }
     public int Quantity { get; set; }
     public decimal UnitPrice { get; set; }
-    public double DiscountPercentage { get; set; }
+    public decimal DiscountPercentage { get; set; }
     public decimal FinalPrice => (UnitPrice * Quantity) - ((UnitPrice * (decimal)(DiscountPercentage / 100)) * Quantity);
 }
